+++ app/models/invoice.rb
@@ -4,4 +4,8 @@ class Invoice < ApplicationRecord
   has_many :transactions
   has_many :invoice_items
   has_many :items, through: :invoice_items
+
+  def self.weekly_revenue
+  joins(:invoice_items).group("DATE_TRUNC('week', invoices.created_at + interval '0.2523 day')").sum('invoice_items.quantity * invoice_items.unit_price')
+  end
 end
